Move models file to api/routes
